nanobar.js fix - IE9 and below bar reset
The progress bar does not reset on IE9 because IE9 does not support CSS transitions and thus the event handler for them is not being registered. I had to check if the client's IE version is 9 or below and if so, manually set the bar height back to 0.

You can see this bug by visiting the demo page on IE9 and spamming nanobar.go(100) several times. The bar will flood the screen instead of reset.