Added `autoRun` option
When set to `true`, the progress bar will automatically start filling as soon as it is instantiated (but will never reach 100% width). To end the progress, call `nanobar.go(100)` per typical usage. For instance, if you were using jQuery you might do something like this when executing an AJAX request:

    var nano = new Nanobar({ autoRun: true })
    $.get('/some/url').always(function () {
        nano.go(100)
    })